ID TRIPULANTE AHORA ES UINT16_T, NO MAS UINT8_T (EN TODO EL TP) PERDON POR LAS MAYUSCULAS

+++ mi-ram-hq/src/comunicacion.c
@@ -27,7 +27,7 @@ void procesar_conexion(void* void_args) {
                 break;
 
             case INICIAR_PATOTA:;
-                uint8_t n_tripulantes;
+                uint16_t n_tripulantes;
                 // char* tareas;
                 t_list* tareas;
                 t_list* posiciones;
@@ -47,7 +47,7 @@ void procesar_conexion(void* void_args) {
 
             case MOVIMIENTO:
             {
-                uint8_t id_tripulante;
+                uint16_t id_tripulante;
                 t_posicion *origen, *destino;
                 if (!recv_movimiento(cliente_socket, &id_tripulante, &origen, &destino)) {
                     log_error(logger, "Error recibiendo movimiento");
@@ -61,7 +61,7 @@ void procesar_conexion(void* void_args) {
             }
             case EXPULSAR_TRIPULANTE:
             {
-                uint8_t id_tripulante;
+                uint16_t id_tripulante;
                  if (recv_tripulante(cliente_socket, &id_tripulante)){
                      int err = expulsar_tripulante(id_tripulante);
                      chequear_errores(err);
